scrapers/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.16.0

# APIs externes
google-api-python-client==2.111.0
tweepy==4.14.0
praw==7.7.1  # Reddit
feedparser==6.0.10  # RSS

# Rate Limiting
slowapi==0.1.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Utils
python-dotenv==1.0.0
